Organization_Management_Service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.0
motor==3.1.1
pydantic==1.10.12
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
pytest==7.4.0